chore: prepare lean cli branch for merge
