--- pkg/cli/declarative_cmd.go
@@ -1,14 +1,0 @@
-package cli
-
-import "github.com/spf13/cobra"
-
-func newDeclarativeCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "declarative",
-		Short: "Manage declarative CUE configuration",
-		Long:  "Generate and work with declarative CUE configuration modules for QuackStack.",
-	}
-
-	cmd.AddCommand(newDeclarativeInitCmd())
-	return cmd
-}
